requirements.txt
datasets 
Pillow 
transformers 
torch
Werkzeug
Flask